ml/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
cairosvg>=2.7.0
albumentations>=1.3.0
h5py>=3.9.0
PyYAML>=6.0
wandb>=0.16.0
onnx>=1.14.0
onnxruntime>=1.16.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
